Added extra error handling
Added extra error handling to provide a error message when a docker container fails to start or create.
Sometimes the interceptor failed to startup due to not being able to start a docker container because the port was already in use.
Now the interceptor will provide a useful error message to the user when this happens.